+++ src/main/java/site/offload/offloadserver/db/repository/MemberRepository.java
@@ -3,5 +3,5 @@ package site.offload.offloadserver.db.repository;
 import org.springframework.data.repository.CrudRepository;
 import site.offload.offloadserver.db.entity.Member;
 
-public interface MemberRepository extends CrudRepository<Member, Integer> {
+public interface MemberRepository extends CrudRepository<Member, Long> {
 }
